+++ src/main/java/org/sopt/exception/GlobalExceptionHandler.java
@@ -1,6 +1,7 @@
 package org.sopt.exception;
 
 import org.sopt.dto.response.ApiResponse;
+import org.sopt.exception.common.BusinessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -8,8 +9,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
-    @ExceptionHandler(PostNotFoundException.class)
-    public ResponseEntity<ApiResponse<Void>> handlePostNotFound(PostNotFoundException e){
+    @ExceptionHandler(BusinessException.class)
+    public ResponseEntity<ApiResponse<Void>> handleBusinessException(BusinessException e){
 
         ErrorCode errorCode = e.getErrorCode();
