+++ src/main/java/org/sopt/exception/common/BusinessException.java
@@ -1,0 +1,17 @@
+package org.sopt.exception.common;
+
+import org.sopt.exception.ErrorCode;
+
+public class BusinessException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
