+++ src/main/java/org/sopt/exception/notfound/NotFoundException.java
@@ -1,0 +1,11 @@
+package org.sopt.exception.notfound;
+
+import org.sopt.exception.ErrorCode;
+import org.sopt.exception.common.BusinessException;
+
+public abstract class NotFoundException extends BusinessException {
+
+    protected NotFoundException(ErrorCode errorCode){
+        super(errorCode);
+    }
+}
