+++ src/main/java/org/sopt/exception/notfound/PostNotFoundException.java
@@ -1,0 +1,10 @@
+package org.sopt.exception.notfound;
+
+import org.sopt.exception.ErrorCode;
+
+public class PostNotFoundException extends NotFoundException {
+
+    public PostNotFoundException() {
+        super(ErrorCode.POST_NOT_FOUND);
+    }
+}
