+++ src/main/java/org/sopt/exception/notfound/UserNotFoundException.java
@@ -1,0 +1,10 @@
+package org.sopt.exception.notfound;
+
+import org.sopt.exception.ErrorCode;
+
+public class UserNotFoundException extends NotFoundException {
+
+    public UserNotFoundException() {
+        super(ErrorCode.USER_NOT_FOUND);
+    }
+}
